+++ app/assets/javascripts/components/story/StoryHistoryLocation.js
@@ -13,20 +13,22 @@ const StoryHistoryLocation = ({ id, url }) =>
       <button
         className="btn btn-default btn-clipboard"
         data-clipboard-target={`#story-link-${id}`}
-        type="button"
+        type="button" 
+        title={I18n.t('story.events.copy_url')}
       >
-        <img src="/clippy.svg" alt="Copy to clipboard" className="clipboard-icon" />
+        <img src="/clippy.svg" alt={I18n.t('story.events.copy_url')} className="clipboard-icon" />
       </button>
 
       <button
         className="btn btn-default btn-clipboard-id btn-clipboard"
         data-clipboard-text={`#${id}`}
         type="button"
+        title={I18n.t('story.events.copy_id')}
       >
         ID
       </button>
 
-      <button className="btn btn-default toggle-history">
+      <button className="btn btn-default toggle-history" title={I18n.t('story.events.view_history')} >
         <i className="mi md-18">history</i>
       </button>